+++ src/main/java/org/harbu/search/problem/Heuristic.java
@@ -2,5 +2,5 @@ package org.harbu.search.problem;
 
 public interface Heuristic<T extends State<T>> {
 
-    public int calculate(T node);
+    public double calculate(T node);
 }
