+++ src/main/java/org/harbu/search/problem/NoHeuristic.java
@@ -3,7 +3,7 @@ package org.harbu.search.problem;
 public class NoHeuristic<T extends State<T>> implements Heuristic<T> {
 
     @Override
-    public int calculate(T node) {
+    public double calculate(T node) {
         return 0;
     }
 }
